feat(app): add routing
