+++ .gitignore
@@ -3,6 +3,7 @@ node_modules
 .DS_Store
 *.log
 .env
+.env.*
 .now
 .now.personal
 .now.team
