+++ packages/app/package.json
@@ -29,7 +29,8 @@
     "isomorphic-unfetch": "^3.0.0",
     "jsonwebtoken": "^8.5.1",
     "node-fetch": "^2.6.0",
-    "sirv-cli": "^0.4.5"
+    "sirv-cli": "^0.4.5",
+    "svelte-routing": "^1.4.2"
   },
   "license": "MIT"
 }
