+++ packages/app/src/pages/index.svelte
@@ -1,0 +1,12 @@
+<script>
+  import Form from '../components/Form/Form.svelte'
+</script>
+
+<slot>
+  <Form />
+  <div class="parks">
+    <!-- {#each parkData.data as park}
+      <ParkCard parkData={park} />
+    {/each} -->
+  </div>
+</slot>
